+++ lib/MIP/Recipes/Analysis/Frequency_annotation.pm
@@ -1,10 +1,11 @@
-package MIP::Recipes::Analysis::RECIPE_NAME;
+package MIP::Recipes::Analysis::Frequency_annotation;
 
 use 5.026;
 use Carp;
 use charnames qw{ :full :short };
 use English qw{ -no_match_vars };
-use File::Spec::Functions qw{ catdir catfile };
+use File::Basename qw{ dirname };
+use File::Spec::Functions qw{ catdir catfile devnull };
 use open qw{ :encoding(UTF-8) :std };
 use Params::Check qw{ allow check last_error };
 use strict;
@@ -17,7 +18,7 @@ use autodie qw{ :all };
 use Readonly;
 
 ## MIPs lib/
-use MIP::Constants qw{ $LOG $NEWLINE $UNDERSCORE };
+use MIP::Constants qw{ $DASH $DOT $LOG $NEWLINE $PIPE $SPACE $UNDERSCORE };
 
 BEGIN {
 
@@ -28,13 +29,13 @@ BEGIN {
     our $VERSION = 1.00;
 
     # Functions and variables which can be optionally exported
-    our @EXPORT_OK = qw{ analysis_RECIPE_NAME };
+    our @EXPORT_OK = qw{ analysis_frequency_annotation };
 
 }
 
-sub analysis_RECIPE_NAME {
+sub analysis_frequency_annotation {
 
-## Function : DESCRIPTION OF RECIPE
+## Function : Annotate vcf with allelle frequencies
 ## Returns  :
 ## Arguments: $active_parameter_href   => Active parameters for this analysis hash {REF}
 ##          : $case_id                 => Family id
@@ -44,7 +45,6 @@ sub analysis_RECIPE_NAME {
 ##          : $parameter_href          => Parameter hash {REF}
 ##          : $profile_base_command    => Submission profile base command
 ##          : $recipe_name             => Recipe name
-##          : $sample_id               => Sample id
 ##          : $sample_info_href        => Info on samples and case hash {REF}
 
     my ($arg_href) = @_;
@@ -56,7 +56,6 @@ sub analysis_RECIPE_NAME {
     my $job_id_href;
     my $parameter_href;
     my $recipe_name;
-    my $sample_id;
     my $sample_info_href;
 
     ## Default(s)
@@ -115,12 +114,6 @@ sub analysis_RECIPE_NAME {
             store       => \$recipe_name,
             strict_type => 1,
         },
-        sample_id => {
-            defined     => 1,
-            required    => 1,
-            store       => \$sample_id,
-            strict_type => 1,
-        },
         sample_info_href => {
             default     => {},
             defined     => 1,
@@ -133,12 +126,14 @@ sub analysis_RECIPE_NAME {
     check( $tmpl, $arg_href, 1 ) or croak q{Could not parse arguments!};
 
     use MIP::Get::File qw{ get_io_files };
-    use MIP::Get::Parameter qw{get_recipe_attributes  get_recipe_resources };
-    use MIP::Program::PATH::TO::PROGRAMS qw{ COMMANDS_SUB };
+    use MIP::Get::Parameter qw{ get_recipe_attributes get_recipe_resources };
     use MIP::Parse::File qw{ parse_io_outfiles };
+    use MIP::Program::Variantcalling::Bcftools
+      qw{ bcftools_concat bcftools_index bcftools_view };
+    use MIP::Program::Variantcalling::Vcfanno qw{ vcfanno };
     use MIP::Processmanagement::Processes qw{ submit_recipe };
-    use MIP::Sample_info
-      qw{ set_recipe_metafile_in_sample_info set_recipe_outfile_in_sample_info };
+    use MIP::Recipes::Analysis::Xargs qw{ xargs_command };
+    use MIP::Sample_info qw{ set_recipe_outfile_in_sample_info };
     use MIP::Script::Setup_script qw{ setup_script };
 
     ### PREPROCESSING:
@@ -150,20 +145,18 @@ sub analysis_RECIPE_NAME {
     ## Get the io infiles per chain and id
     my %io = get_io_files(
         {
-            id             => $sample_id,
+            id             => $case_id,
             file_info_href => $file_info_href,
             parameter_href => $parameter_href,
             recipe_name    => $recipe_name,
             stream         => q{in},
         }
     );
-    my $indir_path_prefix  = $io{in}{dir_path_prefix};
     my $infile_name_prefix = $io{in}{file_name_prefix};
-    my $infile_path        = $io{in}{file_path};
-    my $infile_path_prefix = $io{in}{file_path_prefix};
-    my $infile_suffix      = $io{in}{file_suffix};
+    my %infile_path        = %{ $io{in}{file_path_href} };
 
-    my $job_id_chain = get_recipe_attributes(
+    my @contigs_size_ordered = @{ $file_info_href->{contigs_size_ordered} };
+    my $job_id_chain         = get_recipe_attributes(
         {
             attribute      => q{chain},
             parameter_href => $parameter_href,
@@ -178,36 +171,39 @@ sub analysis_RECIPE_NAME {
         }
     );
 
+    ## Set and get the io files per chain, id and stream
     %io = (
         %io,
         parse_io_outfiles(
             {
-                chain_id               => $job_id_chain,
-                id                     => $sample_id,
-                file_info_href         => $file_info_href,
-                file_name_prefixes_ref => [$infile_name_prefix],
-                outdata_dir            => $active_parameter_href->{outdata_dir},
-                parameter_href         => $parameter_href,
-                recipe_name            => $recipe_name,
+                chain_id         => $job_id_chain,
+                id               => $case_id,
+                file_info_href   => $file_info_href,
+                file_name_prefix => $infile_name_prefix,
+                iterators_ref    => \@contigs_size_ordered,
+                outdata_dir      => $active_parameter_href->{outdata_dir},
+                parameter_href   => $parameter_href,
+                recipe_name      => $recipe_name,
             }
         )
     );
 
-    my $outfile_name_prefix = $io{out}{file_name_prefix};
-    my $outfile_path        = $io{out}{file_path};
+    my @outfile_paths       = @{ $io{out}{file_paths} };
     my $outfile_path_prefix = $io{out}{file_path_prefix};
+    my %outfile_path        = %{ $io{out}{file_path_href} };
     my $outfile_suffix      = $io{out}{file_suffix};
 
     ## Filehandles
     # Create anonymous filehandle
-    my $FILEHANDLE = IO::Handle->new();
+    my $FILEHANDLE      = IO::Handle->new();
+    my $XARGSFILEHANDLE = IO::Handle->new();
 
     ## Creates recipe directories (info & data & script), recipe script filenames and writes sbatch header
     my ( $recipe_file_path, $recipe_info_path ) = setup_script(
         {
             active_parameter_href           => $active_parameter_href,
             core_number                     => $recipe_resource{core_number},
-            directory_id                    => $sample_id,
+            directory_id                    => $case_id,
             FILEHANDLE                      => $FILEHANDLE,
             job_id_href                     => $job_id_href,
             log                             => $log,
@@ -223,9 +219,103 @@ sub analysis_RECIPE_NAME {
 
     say {$FILEHANDLE} q{## } . $recipe_name;
 
-###############################
-###RECIPE TOOL COMMANDS HERE###
-###############################
+    ## Create file commands for xargs
+    my ( $xargs_file_counter, $xargs_file_path_prefix ) = xargs_command(
+        {
+            core_number      => $recipe_resource{core_number},
+            FILEHANDLE       => $FILEHANDLE,
+            file_path        => $recipe_file_path,
+            recipe_info_path => $recipe_info_path,
+            XARGSFILEHANDLE  => $XARGSFILEHANDLE,
+        }
+    );
+
+  CONTIG:
+    foreach my $contig (@contigs_size_ordered) {
+
+        ## Get parameters
+        my $stderrfile_path =
+          $xargs_file_path_prefix . $DOT . $contig . $DOT . q{stderr.txt};
+
+        vcfanno(
+            {
+                FILEHANDLE           => $XARGSFILEHANDLE,
+                infile_path          => $infile_path{$contig},
+                stderrfile_path      => $stderrfile_path,
+                toml_configfile_path => $active_parameter_href->{fqa_vcfanno_config},
+            }
+        );
+
+        print {$XARGSFILEHANDLE} $PIPE . $SPACE;
+
+        bcftools_view(
+            {
+                FILEHANDLE             => $XARGSFILEHANDLE,
+                infile_path            => $DASH,
+                outfile_path           => $outfile_path{$contig},
+                output_type            => q{z},
+                stderrfile_path_append => $stderrfile_path,
+            }
+        );
+        say {$XARGSFILEHANDLE} $NEWLINE;
+
+    }
+
+    say {$FILEHANDLE} q{## Index outfiles};
+
+    ## Create file commands for xargs
+    ( $xargs_file_counter, $xargs_file_path_prefix ) = xargs_command(
+        {
+            core_number        => $recipe_resource{core_number},
+            FILEHANDLE         => $FILEHANDLE,
+            file_path          => $recipe_file_path,
+            recipe_info_path   => $recipe_info_path,
+            XARGSFILEHANDLE    => $XARGSFILEHANDLE,
+            xargs_file_counter => $xargs_file_counter,
+        }
+    );
+
+  CONTIG:
+    foreach my $contig (@contigs_size_ordered) {
+        ## Get parameters
+        my $stderrfile_path =
+          $xargs_file_path_prefix . $DOT . $contig . $DOT . q{stderr.txt};
+
+        bcftools_index(
+            {
+                FILEHANDLE      => $XARGSFILEHANDLE,
+                infile_path     => $outfile_path{$contig},
+                output_type     => q{tbi},
+                stderrfile_path => $stderrfile_path,
+            }
+        );
+        say {$XARGSFILEHANDLE} $NEWLINE;
+    }
+
+    close $XARGSFILEHANDLE or $log->logcroak(q{Could not close XARGSFILEHANDLE});
+
+    say {$FILEHANDLE} q{## Concatenate outfiles};
+
+    bcftools_concat(
+        {
+            FILEHANDLE       => $FILEHANDLE,
+            infile_paths_ref => \@outfile_paths,
+            output_type      => q{z},
+            outfile_path     => $outfile_path_prefix . $DOT . q{vcf.gz},
+            rm_dups          => 0,
+            threads          => $recipe_resource{core_number} - 1,
+        }
+    );
+    say {$FILEHANDLE} $NEWLINE;
+
+    bcftools_index(
+        {
+            infile_path => $outfile_path_prefix . $DOT . q{vcf.gz},
+            FILEHANDLE  => $FILEHANDLE,
+            output_type => q{tbi},
+        }
+    );
+    say {$FILEHANDLE} $NEWLINE;
 
     ## Close FILEHANDLES
     close $FILEHANDLE or $log->logcroak(q{Could not close FILEHANDLE});
@@ -235,42 +325,28 @@ sub analysis_RECIPE_NAME {
         ## Collect QC metadata info for later use
         set_recipe_outfile_in_sample_info(
             {
-                infile           => $outfile_name_prefix,
-                path             => $outfile_path,
+                path             => $outfile_path_prefix . $DOT . q{vcf.gz},
                 recipe_name      => $recipe_name,
-                sample_id        => $sample_id,
                 sample_info_href => $sample_info_href,
             }
         );
 
-        my $most_complete_format_key =
-          q{most_complete} . $UNDERSCORE . substr $outfile_suffix, 1;
-        set_processing_metafile_in_sample_info(
-            {
-                metafile_tag     => $most_complete_format_key,
-                path             => $outfile_path,
-                sample_id        => $sample_id,
-                sample_info_href => $sample_info_href,
-            }
-        );
-
-        ## MODIY THE "dependency_metod" TO HOW YOU WANT SLURM TO PROCESSES UPSTREAM AND DOWNSTREAM DEPENDENCIES
         submit_recipe(
             {
                 base_command            => $profile_base_command,
                 case_id                 => $case_id,
-                dependency_method       => q{sample_to_island},
+                dependency_method       => q{sample_to_case},
                 infile_lane_prefix_href => $infile_lane_prefix_href,
                 job_id_chain            => $job_id_chain,
                 job_id_href             => $job_id_href,
                 log                     => $log,
                 recipe_file_path        => $recipe_file_path,
-                sample_id               => $sample_id,
+                sample_ids_ref          => \@{ $active_parameter_href->{sample_ids} },
                 submission_profile      => $active_parameter_href->{submission_profile},
             }
         );
     }
-    return;
+    return 1;
 }
 
 1;